% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clean_titles}
\alias{clean_titles}
\title{Clean Title Names}
\usage{
clean_titles(data, col = "title")
}
\arguments{
\item{data}{data frame of arXiv records.}

\item{col}{text column in the data frame to clean up}
}
\value{
Character vector of cleaned text.
}
\description{
Standardize a text vector
}
\examples{



}
