% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdf.R
\name{download_pdf}
\alias{download_pdf}
\title{Download PDFs}
\usage{
download_pdf(data, link_col = "link_pdf", directory = ".")
}
\arguments{
\item{data}{data frame of arXiv records}

\item{link_col}{name of column containing the PDF locations. \code{link_pdf} by default.}

\item{directory}{directory to download files to. Current working directory by default.}
}
\value{
None
}
\description{
Takes the results of an arxiv API query and downloads papers to disk.
Papers are saved with their arXiv id, e.g. "2001.12345v1.pdf"
}
\examples{
# get records of papers, then download them
# ml <- get_records("cs.LG", "20200801 TO 20200802", 100)
# download_pdf(ml)

}
