% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_count.R
\name{get_record_count}
\alias{get_record_count}
\title{Get the number of papers}
\usage{
get_record_count(cat, date_range)
}
\arguments{
\item{cat}{\href{https://arxiv.org/category_taxonomy}{arXiv category} name}

\item{date_range}{Submission dates in "YYYYMMDD TO YYYYMMDD" format}
}
\value{
\code{int} of papers meeting the criteria
}
\description{
Queries the arXiv API to get the count of papers submitted to a given category during a stated date range.
}
\examples{
# Get the count of papers submitted to cs.LG on August 1, 2020
ml_count <- get_record_count("cs.LG", "20200801 TO 20200802")

}
