% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get paper metadata records}
\usage{
get_records(cat, date_range, lim)
}
\arguments{
\item{cat}{\href{https://arxiv.org/category_taxonomy}{arXiv category} name}

\item{date_range}{Submission dates in "YYYYMMDD TO YYYYMMDD" format}

\item{lim}{Total number of records to retrieve}
}
\value{
\code{data.frame} of metadata
}
\description{
Queries the arXiv API to get metadata for papers submitted to a given category during a stated date range.
}
\examples{
# Get the count of papers submitted to cs.LG on August 1, 2020
ml <- get_records("cs.LG", "20200801 TO 20200805", 100)

}
