% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get paper metadata records}
\usage{
get_records(cat, date_range, lim = NULL)
}
\arguments{
\item{cat}{\href{https://arxiv.org/category_taxonomy}{arXiv category} name(s) to retrieve papers for. Search for papers listed under multiple categories by using the pipe ("|") delimiter.}

\item{date_range}{Submission dates in "YYYYMMDD TO YYYYMMDD" format}

\item{lim}{Total number of records to retrieve. If the function is called interactively and no lim value is given, the function will check how many records the query will return and prompt the user to proceed.}
}
\value{
\code{data.frame} of metadata
}
\description{
Queries the arXiv API to get metadata for papers submitted to a given category during a specified date range. Records are ordered by earliest submitted.
}
\examples{
# Get the first 10 papers submitted to *both* math.CO and math.PR in 2019
# papers <- get_records("math.CO | math.PR", "2019 TO 2020", 10)

}
